Remove $callback and move method into AjaxAction
