+++ src/Admin/ReenableCssTransientCachingAjaxAction.php
@@ -7,6 +7,9 @@
 
 namespace AmpProject\AmpWP\Admin;
 
+use AMP_Options_Manager;
+use AmpProject\AmpWP\Option;
+
 /**
  * Base class to define a new AJAX action.
  *
@@ -29,13 +32,6 @@ final class ReenableCssTransientCachingAjaxAction {
 	const AJAX_ACTION = 'reenable_css_transient_caching';
 
 	/**
-	 * Callback to execute when an AJAX request was received.
-	 *
-	 * @var callable
-	 */
-	private $callback;
-
-	/**
 	 * Selector to attach the click handler to.
 	 *
 	 * @var string
@@ -45,14 +41,11 @@ final class ReenableCssTransientCachingAjaxAction {
 	/**
 	 * Instantiate a ReenableCssTransientCachingAjaxAction instance.
 	 *
-	 * @param callable $callback Callback function to call when the AJAX request came in.
-	 * @param string   $selector Optional. Selector to attach the click event to. Leave empty to skip the click handler.
+	 * @param string $selector Optional. Selector to attach the click event to. Leave empty to skip the click handler.
 	 */
 	public function __construct(
-		callable $callback,
 		$selector = ''
 	) {
-		$this->callback = $callback;
 		$this->selector = $selector;
 	}
 
@@ -60,8 +53,8 @@ final class ReenableCssTransientCachingAjaxAction {
 	 * Register the AJAX action with the WordPress system.
 	 */
 	public function register() {
-		add_action( 'wp_ajax_' . self::AJAX_ACTION, $this->callback );
 		add_action( static::BACKEND_ENQUEUE_ACTION, [ $this, 'register_ajax_script' ] );
+		add_action( 'wp_ajax_' . self::AJAX_ACTION, [ $this, 'reenable_css_transient_caching' ] );
 	}
 
 	/**
@@ -100,5 +93,24 @@ JS_SCRIPT;
 
 		wp_enqueue_script( 'wp-util' );
 		wp_add_inline_script( 'wp-util', $script );
+	}
+
+	/**
+	 * Re-enable the CSS Transient caching.
+	 *
+	 * This is triggered via an AJAX call from the Site Health panel.
+	 */
+	public function reenable_css_transient_caching() {
+		if ( ! current_user_can( 'manage_options' ) ) {
+			wp_send_json_error( 'Unauthorized.', 401 );
+		}
+
+		$result = AMP_Options_Manager::update_option( Option::DISABLE_CSS_TRANSIENT_CACHING, false );
+
+		if ( false === $result ) {
+			wp_send_json_error( 'CSS transient caching could not be re-enabled.', 500 );
+		}
+
+		wp_send_json_success( 'CSS transient caching was re-enabled.', 200 );
 	}
 }
